[TASK] Move context-related functions to `ContextService`
